+++ src/main/java/ie/cit/cloud/tickets/model/IEventRepository.java
@@ -12,10 +12,10 @@ import java.util.List;
 public interface IEventRepository
 {
 	Performer createPerformer(final String name);
-	Performer getPerformer(final int performerId);
+	Performer getPerformer(final Integer performerId);
 	Performer getPerformer(final String performerName);
 	List<Performer> getPerformers();
-	void deletePerformer(final Long performerId);
+	void deletePerformer(final Integer performerId);
 	
 	Location createLocation(final String locationName, final int ticketCount);
 	Location getLocation(final int locationId);
